Removed CircleMarkScript and fixed some script
CircleMarkScript is changed to MarkScript.

Added a true shuffle function, but it is not 100%, because it needs one old location and a new (if shuffle is pressed, you get a new location). But when both location is the same, it will not work and you have to press it again until it randomizes